Core/PacketIO: Updated packet structures to 9.2.5

+++ src/server/game/Entities/Object/Updates/UpdateFields.cpp
@@ -2998,7 +2998,7 @@ void ActivePlayerData::WriteCreate(ByteBuffer& data, EnumFlag<UpdateFieldFlag> f
     for (uint32 i = 0; i < 12; ++i)
     {
         data << uint32(BuybackPrice[i]);
-        data << uint32(BuybackTimestamp[i]);
+        data << int64(BuybackTimestamp[i]);
     }
     data << uint16(TodayHonorableKills);
     data << uint16(YesterdayHonorableKills);
@@ -4036,7 +4036,7 @@ void ActivePlayerData::WriteUpdate(ByteBuffer& data, Mask const& changesMask, bo
             }
             if (changesMask[605 + i])
             {
-                data << uint32(BuybackTimestamp[i]);
+                data << int64(BuybackTimestamp[i]);
             }
         }
     }
